FEAT: add a new class to represent valid synth state.
This class can be used to safely modify the state from user-settings (as
would be set in PRM files or from the hardware) in a way that is valid.
Because the model is a pydantic class, it should be easy to get LLMs to
generate those models on the fly.